Code reorganization, remove icon_hash_input
